debugging

+++ BusinessApp/UI/CustomerBookTable.cs
@@ -31,7 +31,6 @@ namespace SSC.UI
             this.Size = size;
             this.Location = location;
             customer = c;
-            FillComboBox();
         }
 
         private void InitializeUserControls()
@@ -67,9 +66,6 @@ namespace SSC.UI
                     break;
                 case "orderFood":
                     OpenForm(new CustomerOrderFood(this.Size, this.Location, customer));
-                    break;
-                case "bookTable":
-                    OpenForm(new CustomerBookTable(this.Size, this.Location, customer));
                     break;
                 case "feedback":
                     OpenForm(new CustomerFeedback(this.Size, this.Location, customer));
@@ -116,6 +112,10 @@ namespace SSC.UI
                     comboBox1.Items.Add(table.GetID());
                 }
             }
+            if(comboBox1.Items.Count>0)
+            {
+                comboBox1.SelectedIndex = 0;
+            }
         }
 
         private bool CheckValidations()
@@ -125,7 +125,7 @@ namespace SSC.UI
                 int max = ObjectHandler.GetTableDL().GetTableCapacity(Convert.ToInt32(comboBox1.Text));
                 if (!int.TryParse(guna2TextBox1.Text, out int number) || number < 1 || number > max)
                 {
-                    errorProvider1.SetError(guna2TextBox1, $"Please enter a valid number between 1 and {max}.");
+                    errorProvider1.SetError(guna2TextBox1, $"Please enter a valid number less than or equal to and {max}.");
                     return false;
                 }
                 else
@@ -168,17 +168,8 @@ namespace SSC.UI
                 ObjectHandler.GetReservationDL().SaveReservation(new Reservation(dateTimePicker1.Value,Convert.ToInt32(guna2TextBox1.Text), ObjectHandler.GetCustomerDL().GetCustomerID(customer.GetUsername()), int.Parse(comboBox1.Text)));
                 ObjectHandler.GetTableDL().UpdateTable(new Table(ObjectHandler.GetTableDL().GetTableCapacity(int.Parse(comboBox1.Text)), int.Parse(comboBox1.Text), "Booked"));
                 FillComboBox();
+                guna2TextBox1.Clear();
             }
-        }
-
-        private void label5_Click(object sender, EventArgs e)
-        {
-
-        }
-
-        private void label2_Click(object sender, EventArgs e)
-        {
-
         }
 
         private void viewReservationButton_Click(object sender, EventArgs e)
@@ -205,16 +196,20 @@ namespace SSC.UI
             try
             {
                 int customerID = ObjectHandler.GetCustomerDL().GetCustomerID(customer.GetUsername());
-                if (ObjectHandler.GetReservationDL().GetCustomerReservationCount(customerID) == 1)
+                DialogResult result = MessageBox.Show("Are you sure you want to delete reservation?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (result == DialogResult.Yes)
                 {
-                    ObjectHandler.GetTableDL().UpdateTablesStatus(ObjectHandler.GetTableDL().GetReservedTableID(customerID));
-                    ObjectHandler.GetReservationDL().DeleteReservation(customerID);
-                    MessageBox.Show("Successfully Deleted reservation");
-                    reservationText.Visible = false;
-                }
-                else
-                {
-                    MessageBox.Show("You have no Reservation to delete");
+                    if (ObjectHandler.GetReservationDL().GetCustomerReservationCount(customerID) == 1)
+                    {
+                        ObjectHandler.GetTableDL().UpdateTablesStatus(ObjectHandler.GetTableDL().GetReservedTableID(customerID));
+                        ObjectHandler.GetReservationDL().DeleteReservation(customerID);
+                        MessageBox.Show("Your reservation is deleted", "Confirmation", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        reservationText.Visible = false;
+                    }
+                    else
+                    {
+                        MessageBox.Show("You have no reservation to delete", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
                 }
             }
             catch (Exception ex)
@@ -225,6 +220,8 @@ namespace SSC.UI
 
         private void CustomerBookTable_Load(object sender, EventArgs e)
         {
+            ObjectHandler.GetTableDL().UpdateTablesStatus();
+            FillComboBox();
         }
     }
 }
